make the data vs state vs rendering split cleaner
