+++ martin/src/tui/render.rs
@@ -10,8 +10,9 @@ use ratatui::widgets::canvas::{Canvas, Context, Map, MapResolution, Points};
 use ratatui::widgets::{Block, Paragraph, Row, Sparkline, Table};
 use tracing::Level;
 
+use super::data::Snapshot;
 use super::log::LogLine;
-use super::state::Snapshot;
+use super::state::{LogSize, LogView};
 
 /// A tile asked for this recently is drawn as fresh.
 const FRESH: Duration = Duration::from_secs(10);
@@ -27,28 +28,6 @@ const NOTE_INDENT: &str = "    ";
 const WITH: &str = " with ";
 /// What the pretty format puts between the message of an event and its fields.
 const FIELD_SEPARATOR: &str = ", ";
-
-/// How much of the frame the log pane takes.
-///
-/// The string serialization is the title the log pane wears in that size.
-#[derive(Clone, Copy, Default, PartialEq, Eq, strum::IntoStaticStr)]
-pub enum LogSize {
-    /// The log sits under the panels.
-    #[default]
-    #[strum(serialize = " Log   l expand ")]
-    Normal,
-    /// The log has everything below the header to itself.
-    #[strum(serialize = " Log   l shrink ")]
-    Expanded,
-}
-
-/// How the log pane is being looked at.
-#[derive(Clone, Copy, Default)]
-pub struct LogView {
-    pub size: LogSize,
-    /// How many lines above the newest one the pane stops, `0` while it follows the log.
-    pub scroll: usize,
-}
 
 /// Draws `view` onto the whole frame.
 pub fn frame(frame: &mut Frame, view: &Snapshot, log_view: LogView) {
